fix: block production outputs normalized state diff
