+++ madara/crates/client/block_production/src/lib.rs
@@ -25,7 +25,7 @@ use std::mem;
 use std::sync::Arc;
 use std::time::Instant;
 use tokio::sync::{mpsc, oneshot};
-use util::{state_map_to_state_diff, AdditionalTxInfo, BatchToExecute, BlockExecutionContext, ExecutionStats};
+use util::{AdditionalTxInfo, BatchToExecute, BlockExecutionContext, ExecutionStats};
 
 mod executor;
 pub mod metrics;
@@ -37,7 +37,8 @@ struct PendingBlockState {
     pub transactions: Vec<TransactionWithReceipt>,
     pub events: Vec<EventWithTransactionHash>,
     pub declared_classes: Vec<ConvertedClass>,
-    pub state_diff: StateMaps,
+    /// Unnormalized state diffs.
+    pub state: StateMaps,
 }
 
 impl PendingBlockState {
@@ -52,7 +53,7 @@ impl PendingBlockState {
     pub fn new(header: PendingHeader, initial_state_diffs_storage: HashMap<StorageEntry, Felt>) -> Self {
         Self {
             header,
-            state_diff: StateMaps { storage: initial_state_diffs_storage, ..Default::default() },
+            state: StateMaps { storage: initial_state_diffs_storage, ..Default::default() },
             transactions: vec![],
             events: vec![],
             declared_classes: vec![],
@@ -68,7 +69,7 @@ impl PendingBlockState {
         Ok((
             PendingFullBlock {
                 header: self.header,
-                state_diff: state_map_to_state_diff(backend, &on_top_of_block_id, self.state_diff)
+                state_diff: mc_exec::state_diff::create_normalized_state_diff(backend, &on_top_of_block_id, self.state)
                     .context("Converting state map to state diff")?,
                 transactions: self.transactions,
                 events: self.events,
@@ -154,7 +155,7 @@ impl CurrentPendingState {
                         .cloned()
                         .map(|event| EventWithTransactionHash { event, transaction_hash: converted_tx.hash }),
                 );
-                self.block.state_diff.extend(&state_diff);
+                self.block.state.extend(&state_diff);
                 self.block.transactions.push(TransactionWithReceipt { transaction: converted_tx.transaction, receipt });
             }
         }
@@ -517,7 +518,7 @@ impl BlockProductionTask {
 #[cfg(test)]
 pub(crate) mod tests {
     use crate::BlockProductionStateNotification;
-    use crate::{metrics::BlockProductionMetrics, util::state_map_to_state_diff, BlockProductionTask};
+    use crate::{metrics::BlockProductionMetrics, BlockProductionTask};
     use blockifier::{
         bouncer::{BouncerConfig, BouncerWeights},
         state::cached_state::StateMaps,
@@ -975,7 +976,8 @@ pub(crate) mod tests {
             replaced_classes,
         };
 
-        let mut actual = state_map_to_state_diff(&backend, &Option::<_>::None, state_map).unwrap();
+        let mut actual =
+            mc_exec::state_diff::create_normalized_state_diff(&backend, &Option::<_>::None, state_map).unwrap();
 
         actual.storage_diffs.sort_by(|a, b| a.address.cmp(&b.address));
         actual.storage_diffs.iter_mut().for_each(|s| s.storage_entries.sort_by(|a, b| a.key.cmp(&b.key)));
