+++ madara/crates/client/block_production/src/util.rs
@@ -1,17 +1,13 @@
-use blockifier::{state::cached_state::StateMaps, transaction::transaction_execution::Transaction};
-use mc_db::{db_block_id::DbBlockId, MadaraBackend};
+use blockifier::transaction::transaction_execution::Transaction;
+use mc_db::MadaraBackend;
 use mc_mempool::L1DataProvider;
 use mp_block::header::{BlockTimestamp, GasPrices, PendingHeader};
 use mp_chain_config::{L1DataAvailabilityMode, StarknetVersion};
 use mp_class::ConvertedClass;
-use mp_convert::{Felt, ToFelt};
-use mp_state_update::{
-    ContractStorageDiffItem, DeclaredClassItem, DeployedContractItem, NonceUpdate, ReplacedClassItem, StateDiff,
-    StorageEntry,
-};
-use starknet_api::{core::ContractAddress, StarknetApiError};
+use mp_convert::Felt;
+use starknet_api::StarknetApiError;
 use std::{
-    collections::{hash_map, HashMap, VecDeque},
+    collections::VecDeque,
     ops::{Add, AddAssign},
     sync::Arc,
     time::{Duration, SystemTime},
@@ -154,267 +150,5 @@ pub(crate) fn create_execution_context(
         l1_gas_price: l1_data_provider.get_gas_prices(),
         l1_da_mode: backend.chain_config().l1_da_mode,
         block_n,
-    }
-}
-
-pub(crate) fn state_map_to_state_diff(
-    backend: &MadaraBackend,
-    on_top_of: &Option<DbBlockId>,
-    diff: StateMaps,
-) -> anyhow::Result<StateDiff> {
-    let mut backing_map = HashMap::<ContractAddress, usize>::default();
-    let mut storage_diffs = Vec::<ContractStorageDiffItem>::default();
-    for ((address, key), value) in diff.storage {
-        match backing_map.entry(address) {
-            hash_map::Entry::Vacant(e) => {
-                e.insert(storage_diffs.len());
-                storage_diffs.push(ContractStorageDiffItem {
-                    address: address.to_felt(),
-                    storage_entries: vec![StorageEntry { key: key.to_felt(), value }],
-                });
-            }
-            hash_map::Entry::Occupied(e) => {
-                storage_diffs[*e.get()].storage_entries.push(StorageEntry { key: key.to_felt(), value });
-            }
-        }
-    }
-
-    let mut deprecated_declared_classes = Vec::default();
-    for (class_hash, _) in diff.declared_contracts {
-        if !diff.compiled_class_hashes.contains_key(&class_hash) {
-            deprecated_declared_classes.push(class_hash.to_felt());
-        }
-    }
-
-    let declared_classes = diff
-        .compiled_class_hashes
-        .iter()
-        .map(|(class_hash, compiled_class_hash)| DeclaredClassItem {
-            class_hash: class_hash.to_felt(),
-            compiled_class_hash: compiled_class_hash.to_felt(),
-        })
-        .collect();
-
-    let nonces = diff
-        .nonces
-        .into_iter()
-        .map(|(contract_address, nonce)| NonceUpdate {
-            contract_address: contract_address.to_felt(),
-            nonce: nonce.to_felt(),
-        })
-        .collect();
-
-    let mut deployed_contracts = Vec::new();
-    let mut replaced_classes = Vec::new();
-    for (contract_address, new_class_hash) in diff.class_hashes {
-        let replaced = if let Some(on_top_of) = on_top_of {
-            match backend.get_contract_class_hash_at(on_top_of, &contract_address.to_felt())? {
-                Some(class_hash) => class_hash != new_class_hash.to_felt(),
-                None => false,
-            }
-        } else {
-            // Executing genesis block: nothing being redefined here
-            false
-        };
-        if replaced {
-            replaced_classes.push(ReplacedClassItem {
-                contract_address: contract_address.to_felt(),
-                class_hash: new_class_hash.to_felt(),
-            })
-        } else {
-            deployed_contracts.push(DeployedContractItem {
-                address: contract_address.to_felt(),
-                class_hash: new_class_hash.to_felt(),
-            })
-        }
-    }
-
-    Ok(StateDiff {
-        storage_diffs,
-        deprecated_declared_classes,
-        declared_classes,
-        nonces,
-        deployed_contracts,
-        replaced_classes,
-    })
-}
-
-#[cfg(test)]
-mod test {
-    use blockifier::state::cached_state::StateMaps;
-    use mc_db::MadaraBackend;
-    use mp_chain_config::ChainConfig;
-    use mp_state_update::{
-        ContractStorageDiffItem, DeclaredClassItem, DeployedContractItem, NonceUpdate, StateDiff, StorageEntry,
-    };
-    use starknet_api::core::{ClassHash, CompiledClassHash, Nonce};
-    use starknet_core::types::Felt;
-    use std::{collections::HashMap, sync::Arc};
-
-    #[test]
-    fn state_map_to_state_diff() {
-        let backend = MadaraBackend::open_for_testing(Arc::new(ChainConfig::madara_test()));
-
-        let mut nonces = HashMap::new();
-        nonces.insert(Felt::from_hex_unchecked("1").try_into().unwrap(), Nonce(Felt::from_hex_unchecked("1")));
-        nonces.insert(Felt::from_hex_unchecked("2").try_into().unwrap(), Nonce(Felt::from_hex_unchecked("2")));
-        nonces.insert(Felt::from_hex_unchecked("3").try_into().unwrap(), Nonce(Felt::from_hex_unchecked("3")));
-
-        let mut class_hashes = HashMap::new();
-        class_hashes
-            .insert(Felt::from_hex_unchecked("1").try_into().unwrap(), ClassHash(Felt::from_hex_unchecked("0xc1a551")));
-        class_hashes
-            .insert(Felt::from_hex_unchecked("2").try_into().unwrap(), ClassHash(Felt::from_hex_unchecked("0xc1a552")));
-        class_hashes
-            .insert(Felt::from_hex_unchecked("3").try_into().unwrap(), ClassHash(Felt::from_hex_unchecked("0xc1a553")));
-
-        let mut storage = HashMap::new();
-        storage.insert(
-            (Felt::from_hex_unchecked("1").try_into().unwrap(), Felt::from_hex_unchecked("1").try_into().unwrap()),
-            Felt::from_hex_unchecked("1"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("1").try_into().unwrap(), Felt::from_hex_unchecked("2").try_into().unwrap()),
-            Felt::from_hex_unchecked("2"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("1").try_into().unwrap(), Felt::from_hex_unchecked("3").try_into().unwrap()),
-            Felt::from_hex_unchecked("3"),
-        );
-
-        storage.insert(
-            (Felt::from_hex_unchecked("2").try_into().unwrap(), Felt::from_hex_unchecked("1").try_into().unwrap()),
-            Felt::from_hex_unchecked("1"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("2").try_into().unwrap(), Felt::from_hex_unchecked("2").try_into().unwrap()),
-            Felt::from_hex_unchecked("2"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("2").try_into().unwrap(), Felt::from_hex_unchecked("3").try_into().unwrap()),
-            Felt::from_hex_unchecked("3"),
-        );
-
-        storage.insert(
-            (Felt::from_hex_unchecked("3").try_into().unwrap(), Felt::from_hex_unchecked("1").try_into().unwrap()),
-            Felt::from_hex_unchecked("1"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("3").try_into().unwrap(), Felt::from_hex_unchecked("2").try_into().unwrap()),
-            Felt::from_hex_unchecked("2"),
-        );
-        storage.insert(
-            (Felt::from_hex_unchecked("3").try_into().unwrap(), Felt::from_hex_unchecked("3").try_into().unwrap()),
-            Felt::from_hex_unchecked("3"),
-        );
-
-        let mut compiled_class_hashes = HashMap::new();
-        // "0xc1a553" is marked as deprecated by not having a compiled
-        // class hashe
-        compiled_class_hashes.insert(
-            ClassHash(Felt::from_hex_unchecked("0xc1a551")),
-            CompiledClassHash(Felt::from_hex_unchecked("0x1")),
-        );
-        compiled_class_hashes.insert(
-            ClassHash(Felt::from_hex_unchecked("0xc1a552")),
-            CompiledClassHash(Felt::from_hex_unchecked("0x2")),
-        );
-
-        let mut declared_contracts = HashMap::new();
-        declared_contracts.insert(ClassHash(Felt::from_hex_unchecked("0xc1a551")), true);
-        declared_contracts.insert(ClassHash(Felt::from_hex_unchecked("0xc1a552")), true);
-        declared_contracts.insert(ClassHash(Felt::from_hex_unchecked("0xc1a553")), true);
-
-        let state_map = StateMaps { nonces, class_hashes, storage, compiled_class_hashes, declared_contracts };
-
-        let storage_diffs = vec![
-            ContractStorageDiffItem {
-                address: Felt::from_hex_unchecked("1"),
-                storage_entries: vec![
-                    StorageEntry { key: Felt::from_hex_unchecked("1"), value: Felt::ONE },
-                    StorageEntry { key: Felt::from_hex_unchecked("2"), value: Felt::TWO },
-                    StorageEntry { key: Felt::from_hex_unchecked("3"), value: Felt::THREE },
-                ],
-            },
-            ContractStorageDiffItem {
-                address: Felt::from_hex_unchecked("2"),
-                storage_entries: vec![
-                    StorageEntry { key: Felt::from_hex_unchecked("1"), value: Felt::ONE },
-                    StorageEntry { key: Felt::from_hex_unchecked("2"), value: Felt::TWO },
-                    StorageEntry { key: Felt::from_hex_unchecked("3"), value: Felt::THREE },
-                ],
-            },
-            ContractStorageDiffItem {
-                address: Felt::from_hex_unchecked("3"),
-                storage_entries: vec![
-                    StorageEntry { key: Felt::from_hex_unchecked("1"), value: Felt::ONE },
-                    StorageEntry { key: Felt::from_hex_unchecked("2"), value: Felt::TWO },
-                    StorageEntry { key: Felt::from_hex_unchecked("3"), value: Felt::THREE },
-                ],
-            },
-        ];
-
-        let deprecated_declared_classes = vec![Felt::from_hex_unchecked("0xc1a553")];
-
-        let declared_classes = vec![
-            DeclaredClassItem {
-                class_hash: Felt::from_hex_unchecked("0xc1a551"),
-                compiled_class_hash: Felt::from_hex_unchecked("0x1"),
-            },
-            DeclaredClassItem {
-                class_hash: Felt::from_hex_unchecked("0xc1a552"),
-                compiled_class_hash: Felt::from_hex_unchecked("0x2"),
-            },
-        ];
-
-        let nonces = vec![
-            NonceUpdate { contract_address: Felt::from_hex_unchecked("1"), nonce: Felt::from_hex_unchecked("1") },
-            NonceUpdate { contract_address: Felt::from_hex_unchecked("2"), nonce: Felt::from_hex_unchecked("2") },
-            NonceUpdate { contract_address: Felt::from_hex_unchecked("3"), nonce: Felt::from_hex_unchecked("3") },
-        ];
-
-        let deployed_contracts = vec![
-            DeployedContractItem {
-                address: Felt::from_hex_unchecked("1"),
-                class_hash: Felt::from_hex_unchecked("0xc1a551"),
-            },
-            DeployedContractItem {
-                address: Felt::from_hex_unchecked("2"),
-                class_hash: Felt::from_hex_unchecked("0xc1a552"),
-            },
-            DeployedContractItem {
-                address: Felt::from_hex_unchecked("3"),
-                class_hash: Felt::from_hex_unchecked("0xc1a553"),
-            },
-        ];
-
-        let replaced_classes = vec![];
-
-        let expected = StateDiff {
-            storage_diffs,
-            deprecated_declared_classes,
-            declared_classes,
-            nonces,
-            deployed_contracts,
-            replaced_classes,
-        };
-
-        let mut actual = super::state_map_to_state_diff(&backend, &Option::<_>::None, state_map).unwrap();
-
-        actual.storage_diffs.sort_by(|a, b| a.address.cmp(&b.address));
-        actual.storage_diffs.iter_mut().for_each(|s| s.storage_entries.sort_by(|a, b| a.key.cmp(&b.key)));
-        actual.deprecated_declared_classes.sort();
-        actual.declared_classes.sort_by(|a, b| a.class_hash.cmp(&b.class_hash));
-        actual.nonces.sort_by(|a, b| a.contract_address.cmp(&b.contract_address));
-        actual.deployed_contracts.sort_by(|a, b| a.address.cmp(&b.address));
-        actual.replaced_classes.sort_by(|a, b| a.contract_address.cmp(&b.contract_address));
-
-        assert_eq!(
-            actual,
-            expected,
-            "actual: {}\nexpected: {}",
-            serde_json::to_string_pretty(&actual).unwrap_or_default(),
-            serde_json::to_string_pretty(&expected).unwrap_or_default()
-        );
     }
 }
