+++ madara/crates/client/exec/Cargo.toml
@@ -35,6 +35,8 @@ starknet_api = { workspace = true }
 
 # Other
 
+anyhow = { workspace = true }
+itertools = { workspace = true }
 thiserror = { workspace = true }
 
 
@@ -60,3 +62,4 @@ tracing-subscriber = { workspace = true, features = ["env-filter"] }
 rstest = { workspace = true }
 mc-db = { workspace = true, features = ["testing"] }
 tokio = { workspace = true, features = ["rt-multi-thread"] }
+serde_json = { workspace = true }
