+++ madara/crates/client/exec/src/blockifier_state_adapter.rs
@@ -30,21 +30,18 @@ impl BlockifierStateAdapter {
 // It is however properly handled for transaction validator.
 impl StateReader for BlockifierStateAdapter {
     fn get_storage_at(&self, contract_address: ContractAddress, key: StorageKey) -> StateResult<Felt> {
-        let value = match self.on_top_of_block_id {
-            Some(on_top_of_block_id) => self
-                .backend
-                .get_contract_storage_at(&on_top_of_block_id, &contract_address.to_felt(), &key.to_felt())
-                .map_err(|err| {
-                    StateError::StateReadError(format!(
-                        "Failed to retrieve storage value: on={:?}, contract_address={:#x} key={:#x}: {err:#}",
-                        self.on_top_of_block_id,
-                        contract_address.to_felt(),
-                        key.to_felt(),
-                    ))
-                })?
-                .unwrap_or(Felt::ZERO),
-            None => Felt::ZERO,
-        };
+        let value = self
+            .backend
+            .get_contract_storage_at(&self.on_top_of_block_id, &contract_address.to_felt(), &key.to_felt())
+            .map_err(|err| {
+                StateError::StateReadError(format!(
+                    "Failed to retrieve storage value: on={:?}, contract_address={:#x} key={:#x}: {err:#}",
+                    self.on_top_of_block_id,
+                    contract_address.to_felt(),
+                    key.to_felt(),
+                ))
+            })?
+            .unwrap_or(Felt::ZERO);
 
         tracing::debug!(
             "get_storage_at: on={:?}, contract_address={:#x} key={:#x} => {value:#x}",
@@ -57,20 +54,17 @@ impl StateReader for BlockifierStateAdapter {
     }
 
     fn get_nonce_at(&self, contract_address: ContractAddress) -> StateResult<Nonce> {
-        let value = match self.on_top_of_block_id {
-            Some(on_top_of_block_id) => self
-                .backend
-                .get_contract_nonce_at(&on_top_of_block_id, &contract_address.to_felt())
-                .map_err(|err| {
-                    StateError::StateReadError(format!(
-                        "Failed to retrieve nonce: on={:?}, contract_address={:#x}: {err:#}",
-                        self.on_top_of_block_id,
-                        contract_address.to_felt(),
-                    ))
-                })?
-                .unwrap_or(Felt::ZERO),
-            None => Felt::ZERO,
-        };
+        let value = self
+            .backend
+            .get_contract_nonce_at(&self.on_top_of_block_id, &contract_address.to_felt())
+            .map_err(|err| {
+                StateError::StateReadError(format!(
+                    "Failed to retrieve nonce: on={:?}, contract_address={:#x}: {err:#}",
+                    self.on_top_of_block_id,
+                    contract_address.to_felt(),
+                ))
+            })?
+            .unwrap_or(Felt::ZERO);
 
         tracing::debug!(
             "get_nonce_at: on={:?}, contract_address={:#x} => {value:#x}",
@@ -83,20 +77,17 @@ impl StateReader for BlockifierStateAdapter {
 
     /// Blockifier expects us to return 0x0 if the contract is not deployed.
     fn get_class_hash_at(&self, contract_address: ContractAddress) -> StateResult<ClassHash> {
-        let value = match self.on_top_of_block_id {
-            Some(on_top_of_block_id) => self
-                .backend
-                .get_contract_class_hash_at(&on_top_of_block_id, &contract_address.to_felt())
-                .map_err(|err| {
-                    StateError::StateReadError(format!(
-                        "Failed to retrieve class_hash: on={:?}, contract_address={:#x}: {err:#}",
-                        self.on_top_of_block_id,
-                        contract_address.to_felt(),
-                    ))
-                })?
-                .unwrap_or(Felt::ZERO),
-            None => Felt::ZERO,
-        };
+        let value = self
+            .backend
+            .get_contract_class_hash_at(&self.on_top_of_block_id, &contract_address.to_felt())
+            .map_err(|err| {
+                StateError::StateReadError(format!(
+                    "Failed to retrieve class_hash: on={:?}, contract_address={:#x}: {err:#}",
+                    self.on_top_of_block_id,
+                    contract_address.to_felt(),
+                ))
+            })?
+            .unwrap_or(Felt::ZERO);
 
         tracing::debug!(
             "get_class_hash_at: on={:?}, contract_address={:#x} => {value:#x}",
@@ -108,18 +99,14 @@ impl StateReader for BlockifierStateAdapter {
     }
 
     fn get_compiled_contract_class(&self, class_hash: ClassHash) -> StateResult<ContractClass> {
-        let value = match self.on_top_of_block_id {
-            Some(on_top_of_block_id) => {
-                self.backend.get_converted_class(&on_top_of_block_id, &class_hash.to_felt()).map_err(|err| {
-                    StateError::StateReadError(format!(
-                        "Failed to retrieve class_hash: on={:?}, class_hash={:#x}: {err:#}",
-                        self.on_top_of_block_id,
-                        class_hash.to_felt(),
-                    ))
-                })?
-            }
-            None => None,
-        };
+        let value =
+            self.backend.get_converted_class(&self.on_top_of_block_id, &class_hash.to_felt()).map_err(|err| {
+                StateError::StateReadError(format!(
+                    "Failed to retrieve class_hash: on={:?}, class_hash={:#x}: {err:#}",
+                    self.on_top_of_block_id,
+                    class_hash.to_felt(),
+                ))
+            })?;
 
         let value = value.ok_or(StateError::UndeclaredClassHash(class_hash))?;
 
@@ -135,18 +122,13 @@ impl StateReader for BlockifierStateAdapter {
     }
 
     fn get_compiled_class_hash(&self, class_hash: ClassHash) -> StateResult<CompiledClassHash> {
-        let value = match self.on_top_of_block_id {
-            Some(on_top_of_block_id) => {
-                self.backend.get_class_info(&on_top_of_block_id, &class_hash.to_felt()).map_err(|err| {
-                    StateError::StateReadError(format!(
-                        "Failed to retrieve class_hash: on={:?}, class_hash={:#x}: {err:#}",
-                        self.on_top_of_block_id,
-                        class_hash.to_felt(),
-                    ))
-                })?
-            }
-            None => None,
-        };
+        let value = self.backend.get_class_info(&self.on_top_of_block_id, &class_hash.to_felt()).map_err(|err| {
+            StateError::StateReadError(format!(
+                "Failed to retrieve class_hash: on={:?}, class_hash={:#x}: {err:#}",
+                self.on_top_of_block_id,
+                class_hash.to_felt(),
+            ))
+        })?;
 
         let value = value.and_then(|c| c.compiled_class_hash()).ok_or_else(|| {
             StateError::StateReadError(format!(
